p4i integration Integrate 276438 from 4.3 to main: Fixes:    Fixes qmax(qreal,double) build error on ARM RevBy:    Thiago Details:  Fixes qmax(qreal,double) build error on ARM
[git-p4: depot-paths = "//depot/qt/main/": change = 276706]

+++ examples/graphicsview/portedcanvas/canvas.cpp
@@ -423,7 +423,7 @@ void Main::enlarge()
 
 void Main::shrink()
 {
-    canvas.setSceneRect(0, 0, qMax(canvas.width()*3/4, 1.0), qMax(canvas.height()*3/4, 1.0));
+    canvas.setSceneRect(0, 0, qMax(canvas.width()*3/4, qreal(1.0)), qMax(canvas.height()*3/4, qreal(1.0)));
 }
 
 void Main::rotateClockwise()
